// src/shared/types/astro/astro-houses.types.ts
export enum ASTRO_HOUSE_ACTIVITY {
  EMPTY = 'empty',
  ACTIVE = 'active',
}

export enum HOUSE_SYSTEM {
  Placidus = 'Placidus',
  Koch = 'Koch',
  Porphyry = 'Porphyry',
  Regiomontanus = 'Regiomontanus',
  Campanus = 'Campanus',
  EqualMC = 'EqualMC', // равнодомная от MC
  EqualAsc = 'EqualAsc', // равнодомная от Asc
}